+++ src/modules/crm/infrastructure/container.ts
@@ -1,5 +1,7 @@
+import { organizationQueries } from "@/modules/organizations";
 import { makeCrmCommands } from "../application/commands";
 import { makeCrmQueries } from "../application/queries";
+import { makeCustomerDirectory } from "../application/customer-directory";
 import { PrismaCustomerRepository } from "./customer.repository";
 import { PrismaFollowerRepository } from "./follower.repository";
 
@@ -9,3 +11,9 @@ const followers = new PrismaFollowerRepository();
 /** Composition root for the crm module. */
 export const crmCommands = makeCrmCommands({ followers });
 export const crmQueries = makeCrmQueries({ customers, followers });
+
+export const customerDirectory = makeCustomerDirectory({
+  organizations: organizationQueries,
+  customers,
+  followers,
+});